swap analysis backend from cleave to litmus

+++ cmd/hopper/hashcache.go
@@ -163,7 +163,7 @@ func (c *hashCache) close() {
 // fileStat returns the device and inode numbers from an os.FileInfo via Stat_t.
 func fileStat(info os.FileInfo) (dev, inode uint64) {
 	if st, ok := info.Sys().(*syscall.Stat_t); ok {
-		return st.Dev, st.Ino
+		return uint64(st.Dev), st.Ino
 	}
 	return 0, 0
 }
